FXHash post-merge cleanup
