+++ src/ingestors/fxhash/abi.ts
@@ -1,127 +1,6 @@
 export const FXHASH_BASE_FRAME_ABI = [
   {
     inputs: [
-      { internalType: 'address', name: '_owner', type: 'address' },
-      { internalType: 'address', name: '_controller', type: 'address' },
-    ],
-    stateMutability: 'nonpayable',
-    type: 'constructor',
-  },
-  { inputs: [], name: 'AddressZero', type: 'error' },
-  { inputs: [], name: 'Ended', type: 'error' },
-  { inputs: [], name: 'InsufficientFunds', type: 'error' },
-  { inputs: [], name: 'InvalidAllocation', type: 'error' },
-  { inputs: [], name: 'InvalidPayment', type: 'error' },
-  { inputs: [], name: 'InvalidReserve', type: 'error' },
-  { inputs: [], name: 'InvalidToken', type: 'error' },
-  { inputs: [], name: 'MaxAmountExceeded', type: 'error' },
-  { inputs: [], name: 'NewOwnerIsZeroAddress', type: 'error' },
-  { inputs: [], name: 'NoHandoverRequest', type: 'error' },
-  { inputs: [], name: 'NotStarted', type: 'error' },
-  { inputs: [], name: 'TooMany', type: 'error' },
-  { inputs: [], name: 'Unauthorized', type: 'error' },
-  { inputs: [], name: 'ZeroAddress', type: 'error' },
-  {
-    anonymous: false,
-    inputs: [
-      { indexed: false, internalType: 'address', name: '_prevController', type: 'address' },
-      { indexed: false, internalType: 'address', name: '_newController', type: 'address' },
-    ],
-    name: 'ControllerSet',
-    type: 'event',
-  },
-  {
-    anonymous: false,
-    inputs: [
-      { indexed: true, internalType: 'address', name: '_token', type: 'address' },
-      { indexed: true, internalType: 'address', name: '_to', type: 'address' },
-      { indexed: true, internalType: 'uint256', name: '_fid', type: 'uint256' },
-    ],
-    name: 'FrameMinted',
-    type: 'event',
-  },
-  {
-    anonymous: false,
-    inputs: [
-      { indexed: true, internalType: 'address', name: '_token', type: 'address' },
-      { indexed: true, internalType: 'uint256', name: '_reserveId', type: 'uint256' },
-      { indexed: false, internalType: 'uint256', name: '_price', type: 'uint256' },
-      {
-        components: [
-          { internalType: 'uint64', name: 'startTime', type: 'uint64' },
-          { internalType: 'uint64', name: 'endTime', type: 'uint64' },
-          { internalType: 'uint128', name: 'allocation', type: 'uint128' },
-        ],
-        indexed: false,
-        internalType: 'struct ReserveInfo',
-        name: '_reserveInfo',
-        type: 'tuple',
-      },
-      { indexed: false, internalType: 'bool', name: '_openEdition', type: 'bool' },
-      { indexed: false, internalType: 'bool', name: '_timeUnlimited', type: 'bool' },
-      { indexed: false, internalType: 'uint256', name: '_maxAmountPerFid', type: 'uint256' },
-    ],
-    name: 'MintDetailsSet',
-    type: 'event',
-  },
-  {
-    anonymous: false,
-    inputs: [{ indexed: true, internalType: 'address', name: 'pendingOwner', type: 'address' }],
-    name: 'OwnershipHandoverCanceled',
-    type: 'event',
-  },
-  {
-    anonymous: false,
-    inputs: [{ indexed: true, internalType: 'address', name: 'pendingOwner', type: 'address' }],
-    name: 'OwnershipHandoverRequested',
-    type: 'event',
-  },
-  {
-    anonymous: false,
-    inputs: [
-      { indexed: true, internalType: 'address', name: 'oldOwner', type: 'address' },
-      { indexed: true, internalType: 'address', name: 'newOwner', type: 'address' },
-    ],
-    name: 'OwnershipTransferred',
-    type: 'event',
-  },
-  {
-    anonymous: false,
-    inputs: [{ indexed: false, internalType: 'address', name: 'account', type: 'address' }],
-    name: 'Paused',
-    type: 'event',
-  },
-  {
-    anonymous: false,
-    inputs: [
-      { indexed: true, internalType: 'address', name: '_token', type: 'address' },
-      { indexed: true, internalType: 'uint256', name: '_reserveId', type: 'uint256' },
-      { indexed: true, internalType: 'address', name: '_buyer', type: 'address' },
-      { indexed: false, internalType: 'uint256', name: '_amount', type: 'uint256' },
-      { indexed: false, internalType: 'address', name: '_to', type: 'address' },
-      { indexed: false, internalType: 'uint256', name: '_price', type: 'uint256' },
-    ],
-    name: 'Purchase',
-    type: 'event',
-  },
-  {
-    anonymous: false,
-    inputs: [{ indexed: false, internalType: 'address', name: 'account', type: 'address' }],
-    name: 'Unpaused',
-    type: 'event',
-  },
-  {
-    anonymous: false,
-    inputs: [
-      { indexed: true, internalType: 'address', name: '_token', type: 'address' },
-      { indexed: true, internalType: 'address', name: '_creator', type: 'address' },
-      { indexed: false, internalType: 'uint256', name: '_proceeds', type: 'uint256' },
-    ],
-    name: 'Withdrawn',
-    type: 'event',
-  },
-  {
-    inputs: [
       { internalType: 'address', name: '_token', type: 'address' },
       { internalType: 'uint256', name: '_reserveId', type: 'uint256' },
       { internalType: 'uint256', name: '_amount', type: 'uint256' },
@@ -130,83 +9,6 @@ export const FXHASH_BASE_FRAME_ABI = [
     name: 'buy',
     outputs: [],
     stateMutability: 'payable',
-    type: 'function',
-  },
-  { inputs: [], name: 'cancelOwnershipHandover', outputs: [], stateMutability: 'payable', type: 'function' },
-  {
-    inputs: [{ internalType: 'address', name: 'pendingOwner', type: 'address' }],
-    name: 'completeOwnershipHandover',
-    outputs: [],
-    stateMutability: 'payable',
-    type: 'function',
-  },
-  {
-    inputs: [],
-    name: 'controller',
-    outputs: [{ internalType: 'address', name: '', type: 'address' }],
-    stateMutability: 'view',
-    type: 'function',
-  },
-  {
-    inputs: [{ internalType: 'address', name: '_token', type: 'address' }],
-    name: 'getFirstValidReserve',
-    outputs: [{ internalType: 'uint256', name: '', type: 'uint256' }],
-    stateMutability: 'view',
-    type: 'function',
-  },
-  {
-    inputs: [{ internalType: 'address', name: '_token', type: 'address' }],
-    name: 'getLatestUpdate',
-    outputs: [{ internalType: 'uint40', name: '', type: 'uint40' }],
-    stateMutability: 'view',
-    type: 'function',
-  },
-  {
-    inputs: [{ internalType: 'address', name: '_token', type: 'address' }],
-    name: 'getSaleProceeds',
-    outputs: [{ internalType: 'uint128', name: '', type: 'uint128' }],
-    stateMutability: 'view',
-    type: 'function',
-  },
-  {
-    inputs: [{ internalType: 'address', name: '', type: 'address' }],
-    name: 'maxAmounts',
-    outputs: [{ internalType: 'uint256', name: '', type: 'uint256' }],
-    stateMutability: 'view',
-    type: 'function',
-  },
-  {
-    inputs: [
-      { internalType: 'address', name: '_token', type: 'address' },
-      { internalType: 'uint256', name: '_reserveId', type: 'uint256' },
-      { internalType: 'uint256', name: '_fId', type: 'uint256' },
-      { internalType: 'address', name: '_to', type: 'address' },
-    ],
-    name: 'mint',
-    outputs: [],
-    stateMutability: 'nonpayable',
-    type: 'function',
-  },
-  {
-    inputs: [],
-    name: 'owner',
-    outputs: [{ internalType: 'address', name: 'result', type: 'address' }],
-    stateMutability: 'view',
-    type: 'function',
-  },
-  {
-    inputs: [{ internalType: 'address', name: 'pendingOwner', type: 'address' }],
-    name: 'ownershipHandoverExpiresAt',
-    outputs: [{ internalType: 'uint256', name: 'result', type: 'uint256' }],
-    stateMutability: 'view',
-    type: 'function',
-  },
-  { inputs: [], name: 'pause', outputs: [], stateMutability: 'nonpayable', type: 'function' },
-  {
-    inputs: [],
-    name: 'paused',
-    outputs: [{ internalType: 'bool', name: '', type: 'bool' }],
-    stateMutability: 'view',
     type: 'function',
   },
   {
@@ -218,204 +20,10 @@ export const FXHASH_BASE_FRAME_ABI = [
     outputs: [{ internalType: 'uint256', name: '', type: 'uint256' }],
     stateMutability: 'view',
     type: 'function',
-  },
-  { inputs: [], name: 'renounceOwnership', outputs: [], stateMutability: 'payable', type: 'function' },
-  { inputs: [], name: 'requestOwnershipHandover', outputs: [], stateMutability: 'payable', type: 'function' },
-  {
-    inputs: [
-      { internalType: 'address', name: '', type: 'address' },
-      { internalType: 'uint256', name: '', type: 'uint256' },
-    ],
-    name: 'reserves',
-    outputs: [
-      { internalType: 'uint64', name: 'startTime', type: 'uint64' },
-      { internalType: 'uint64', name: 'endTime', type: 'uint64' },
-      { internalType: 'uint128', name: 'allocation', type: 'uint128' },
-    ],
-    stateMutability: 'view',
-    type: 'function',
-  },
-  {
-    inputs: [{ internalType: 'address', name: '_controller', type: 'address' }],
-    name: 'setController',
-    outputs: [],
-    stateMutability: 'nonpayable',
-    type: 'function',
-  },
-  {
-    inputs: [
-      {
-        components: [
-          { internalType: 'uint64', name: 'startTime', type: 'uint64' },
-          { internalType: 'uint64', name: 'endTime', type: 'uint64' },
-          { internalType: 'uint128', name: 'allocation', type: 'uint128' },
-        ],
-        internalType: 'struct ReserveInfo',
-        name: '_reserve',
-        type: 'tuple',
-      },
-      { internalType: 'bytes', name: '_mintDetails', type: 'bytes' },
-    ],
-    name: 'setMintDetails',
-    outputs: [],
-    stateMutability: 'nonpayable',
-    type: 'function',
-  },
-  {
-    inputs: [
-      { internalType: 'uint256', name: '', type: 'uint256' },
-      { internalType: 'address', name: '', type: 'address' },
-    ],
-    name: 'totalMinted',
-    outputs: [{ internalType: 'uint256', name: '', type: 'uint256' }],
-    stateMutability: 'view',
-    type: 'function',
-  },
-  {
-    inputs: [{ internalType: 'address', name: 'newOwner', type: 'address' }],
-    name: 'transferOwnership',
-    outputs: [],
-    stateMutability: 'payable',
-    type: 'function',
-  },
-  { inputs: [], name: 'unpause', outputs: [], stateMutability: 'nonpayable', type: 'function' },
-  {
-    inputs: [{ internalType: 'address', name: '_token', type: 'address' }],
-    name: 'withdraw',
-    outputs: [],
-    stateMutability: 'nonpayable',
-    type: 'function',
-  },
+  }
 ];
 
 export const FXHASH_BASE_FIXED_PRICE_ABI = [
-  { inputs: [], name: 'AddressZero', type: 'error' },
-  { inputs: [], name: 'Ended', type: 'error' },
-  { inputs: [], name: 'InsufficientFunds', type: 'error' },
-  { inputs: [], name: 'InvalidAllocation', type: 'error' },
-  { inputs: [], name: 'InvalidPayment', type: 'error' },
-  { inputs: [], name: 'InvalidProof', type: 'error' },
-  { inputs: [], name: 'InvalidReserve', type: 'error' },
-  { inputs: [], name: 'InvalidShortString', type: 'error' },
-  { inputs: [], name: 'InvalidSignature', type: 'error' },
-  { inputs: [], name: 'InvalidTimes', type: 'error' },
-  { inputs: [], name: 'InvalidToken', type: 'error' },
-  { inputs: [], name: 'NewOwnerIsZeroAddress', type: 'error' },
-  { inputs: [], name: 'NoAllowlist', type: 'error' },
-  { inputs: [], name: 'NoHandoverRequest', type: 'error' },
-  { inputs: [], name: 'NoPublicMint', type: 'error' },
-  { inputs: [], name: 'NoSigningAuthority', type: 'error' },
-  { inputs: [], name: 'NotStarted', type: 'error' },
-  { inputs: [], name: 'OnlyAuthorityOrAllowlist', type: 'error' },
-  { inputs: [], name: 'PassAlreadyClaimed', type: 'error' },
-  { inputs: [], name: 'SlotAlreadyClaimed', type: 'error' },
-  { inputs: [{ internalType: 'string', name: 'str', type: 'string' }], name: 'StringTooLong', type: 'error' },
-  { inputs: [], name: 'TooMany', type: 'error' },
-  { inputs: [], name: 'Unauthorized', type: 'error' },
-  { anonymous: false, inputs: [], name: 'EIP712DomainChanged', type: 'event' },
-  {
-    anonymous: false,
-    inputs: [
-      { indexed: true, internalType: 'address', name: '_token', type: 'address' },
-      { indexed: true, internalType: 'uint256', name: '_reserveId', type: 'uint256' },
-      { indexed: false, internalType: 'uint256', name: '_price', type: 'uint256' },
-      {
-        components: [
-          { internalType: 'uint64', name: 'startTime', type: 'uint64' },
-          { internalType: 'uint64', name: 'endTime', type: 'uint64' },
-          { internalType: 'uint128', name: 'allocation', type: 'uint128' },
-        ],
-        indexed: false,
-        internalType: 'struct ReserveInfo',
-        name: '_reserveInfo',
-        type: 'tuple',
-      },
-      { indexed: false, internalType: 'bytes32', name: '_merkleRoot', type: 'bytes32' },
-      { indexed: false, internalType: 'address', name: '_mintPassSigner', type: 'address' },
-      { indexed: false, internalType: 'bool', name: '_openEdition', type: 'bool' },
-      { indexed: false, internalType: 'bool', name: '_timeUnlimited', type: 'bool' },
-    ],
-    name: 'MintDetailsSet',
-    type: 'event',
-  },
-  {
-    anonymous: false,
-    inputs: [{ indexed: true, internalType: 'address', name: 'pendingOwner', type: 'address' }],
-    name: 'OwnershipHandoverCanceled',
-    type: 'event',
-  },
-  {
-    anonymous: false,
-    inputs: [{ indexed: true, internalType: 'address', name: 'pendingOwner', type: 'address' }],
-    name: 'OwnershipHandoverRequested',
-    type: 'event',
-  },
-  {
-    anonymous: false,
-    inputs: [
-      { indexed: true, internalType: 'address', name: 'oldOwner', type: 'address' },
-      { indexed: true, internalType: 'address', name: 'newOwner', type: 'address' },
-    ],
-    name: 'OwnershipTransferred',
-    type: 'event',
-  },
-  {
-    anonymous: false,
-    inputs: [
-      { indexed: true, internalType: 'address', name: '_token', type: 'address' },
-      { indexed: true, internalType: 'uint256', name: '_reserveId', type: 'uint256' },
-      { indexed: true, internalType: 'address', name: '_claimer', type: 'address' },
-      { indexed: false, internalType: 'uint256', name: '_index', type: 'uint256' },
-    ],
-    name: 'PassClaimed',
-    type: 'event',
-  },
-  {
-    anonymous: false,
-    inputs: [{ indexed: false, internalType: 'address', name: 'account', type: 'address' }],
-    name: 'Paused',
-    type: 'event',
-  },
-  {
-    anonymous: false,
-    inputs: [
-      { indexed: true, internalType: 'address', name: '_token', type: 'address' },
-      { indexed: true, internalType: 'uint256', name: '_reserveId', type: 'uint256' },
-      { indexed: true, internalType: 'address', name: '_buyer', type: 'address' },
-      { indexed: false, internalType: 'uint256', name: '_amount', type: 'uint256' },
-      { indexed: false, internalType: 'address', name: '_to', type: 'address' },
-      { indexed: false, internalType: 'uint256', name: '_price', type: 'uint256' },
-    ],
-    name: 'Purchase',
-    type: 'event',
-  },
-  {
-    anonymous: false,
-    inputs: [
-      { indexed: true, internalType: 'address', name: '_token', type: 'address' },
-      { indexed: true, internalType: 'uint256', name: '_reserveId', type: 'uint256' },
-      { indexed: true, internalType: 'address', name: '_claimer', type: 'address' },
-      { indexed: false, internalType: 'uint256', name: '_index', type: 'uint256' },
-    ],
-    name: 'SlotClaimed',
-    type: 'event',
-  },
-  {
-    anonymous: false,
-    inputs: [{ indexed: false, internalType: 'address', name: 'account', type: 'address' }],
-    name: 'Unpaused',
-    type: 'event',
-  },
-  {
-    anonymous: false,
-    inputs: [
-      { indexed: true, internalType: 'address', name: '_token', type: 'address' },
-      { indexed: true, internalType: 'address', name: '_creator', type: 'address' },
-      { indexed: false, internalType: 'uint256', name: '_proceeds', type: 'uint256' },
-    ],
-    name: 'Withdrawn',
-    type: 'event',
-  },
   {
     inputs: [
       { internalType: 'address', name: '_token', type: 'address' },
@@ -430,122 +38,6 @@ export const FXHASH_BASE_FIXED_PRICE_ABI = [
   },
   {
     inputs: [
-      { internalType: 'address', name: '_token', type: 'address' },
-      { internalType: 'uint256', name: '_reserveId', type: 'uint256' },
-      { internalType: 'address', name: '_to', type: 'address' },
-      { internalType: 'uint256[]', name: '_indexes', type: 'uint256[]' },
-      { internalType: 'bytes32[][]', name: '_proofs', type: 'bytes32[][]' },
-    ],
-    name: 'buyAllowlist',
-    outputs: [],
-    stateMutability: 'payable',
-    type: 'function',
-  },
-  {
-    inputs: [
-      { internalType: 'address', name: '_token', type: 'address' },
-      { internalType: 'uint256', name: '_reserveId', type: 'uint256' },
-      { internalType: 'uint256', name: '_amount', type: 'uint256' },
-      { internalType: 'address', name: '_to', type: 'address' },
-      { internalType: 'uint256', name: '_index', type: 'uint256' },
-      { internalType: 'bytes', name: '_signature', type: 'bytes' },
-    ],
-    name: 'buyMintPass',
-    outputs: [],
-    stateMutability: 'payable',
-    type: 'function',
-  },
-  { inputs: [], name: 'cancelOwnershipHandover', outputs: [], stateMutability: 'payable', type: 'function' },
-  {
-    inputs: [{ internalType: 'address', name: 'pendingOwner', type: 'address' }],
-    name: 'completeOwnershipHandover',
-    outputs: [],
-    stateMutability: 'payable',
-    type: 'function',
-  },
-  {
-    inputs: [],
-    name: 'eip712Domain',
-    outputs: [
-      { internalType: 'bytes1', name: 'fields', type: 'bytes1' },
-      { internalType: 'string', name: 'name', type: 'string' },
-      { internalType: 'string', name: 'version', type: 'string' },
-      { internalType: 'uint256', name: 'chainId', type: 'uint256' },
-      { internalType: 'address', name: 'verifyingContract', type: 'address' },
-      { internalType: 'bytes32', name: 'salt', type: 'bytes32' },
-      { internalType: 'uint256[]', name: 'extensions', type: 'uint256[]' },
-    ],
-    stateMutability: 'view',
-    type: 'function',
-  },
-  {
-    inputs: [
-      { internalType: 'address', name: '_token', type: 'address' },
-      { internalType: 'uint256', name: '_reserveId', type: 'uint256' },
-      { internalType: 'uint256', name: '_reserveNonce', type: 'uint256' },
-      { internalType: 'uint256', name: '_index', type: 'uint256' },
-      { internalType: 'address', name: '_claimer', type: 'address' },
-    ],
-    name: 'generateTypedDataHash',
-    outputs: [{ internalType: 'bytes32', name: '', type: 'bytes32' }],
-    stateMutability: 'view',
-    type: 'function',
-  },
-  {
-    inputs: [{ internalType: 'address', name: '_token', type: 'address' }],
-    name: 'getFirstValidReserve',
-    outputs: [{ internalType: 'uint256', name: '', type: 'uint256' }],
-    stateMutability: 'view',
-    type: 'function',
-  },
-  {
-    inputs: [{ internalType: 'address', name: '_token', type: 'address' }],
-    name: 'getLatestUpdate',
-    outputs: [{ internalType: 'uint40', name: '', type: 'uint40' }],
-    stateMutability: 'view',
-    type: 'function',
-  },
-  {
-    inputs: [{ internalType: 'address', name: '_token', type: 'address' }],
-    name: 'getSaleProceed',
-    outputs: [{ internalType: 'uint128', name: '', type: 'uint128' }],
-    stateMutability: 'view',
-    type: 'function',
-  },
-  {
-    inputs: [
-      { internalType: 'address', name: '', type: 'address' },
-      { internalType: 'uint256', name: '', type: 'uint256' },
-    ],
-    name: 'merkleRoots',
-    outputs: [{ internalType: 'bytes32', name: '', type: 'bytes32' }],
-    stateMutability: 'view',
-    type: 'function',
-  },
-  {
-    inputs: [],
-    name: 'owner',
-    outputs: [{ internalType: 'address', name: 'result', type: 'address' }],
-    stateMutability: 'view',
-    type: 'function',
-  },
-  {
-    inputs: [{ internalType: 'address', name: 'pendingOwner', type: 'address' }],
-    name: 'ownershipHandoverExpiresAt',
-    outputs: [{ internalType: 'uint256', name: 'result', type: 'uint256' }],
-    stateMutability: 'view',
-    type: 'function',
-  },
-  { inputs: [], name: 'pause', outputs: [], stateMutability: 'nonpayable', type: 'function' },
-  {
-    inputs: [],
-    name: 'paused',
-    outputs: [{ internalType: 'bool', name: '', type: 'bool' }],
-    stateMutability: 'view',
-    type: 'function',
-  },
-  {
-    inputs: [
       { internalType: 'address', name: '', type: 'address' },
       { internalType: 'uint256', name: '', type: 'uint256' },
     ],
@@ -553,75 +45,5 @@ export const FXHASH_BASE_FIXED_PRICE_ABI = [
     outputs: [{ internalType: 'uint256', name: '', type: 'uint256' }],
     stateMutability: 'view',
     type: 'function',
-  },
-  { inputs: [], name: 'renounceOwnership', outputs: [], stateMutability: 'payable', type: 'function' },
-  { inputs: [], name: 'requestOwnershipHandover', outputs: [], stateMutability: 'payable', type: 'function' },
-  {
-    inputs: [
-      { internalType: 'address', name: '', type: 'address' },
-      { internalType: 'uint256', name: '', type: 'uint256' },
-    ],
-    name: 'reserveNonce',
-    outputs: [{ internalType: 'uint256', name: '', type: 'uint256' }],
-    stateMutability: 'view',
-    type: 'function',
-  },
-  {
-    inputs: [
-      { internalType: 'address', name: '', type: 'address' },
-      { internalType: 'uint256', name: '', type: 'uint256' },
-    ],
-    name: 'reserves',
-    outputs: [
-      { internalType: 'uint64', name: 'startTime', type: 'uint64' },
-      { internalType: 'uint64', name: 'endTime', type: 'uint64' },
-      { internalType: 'uint128', name: 'allocation', type: 'uint128' },
-    ],
-    stateMutability: 'view',
-    type: 'function',
-  },
-  {
-    inputs: [
-      {
-        components: [
-          { internalType: 'uint64', name: 'startTime', type: 'uint64' },
-          { internalType: 'uint64', name: 'endTime', type: 'uint64' },
-          { internalType: 'uint128', name: 'allocation', type: 'uint128' },
-        ],
-        internalType: 'struct ReserveInfo',
-        name: '_reserve',
-        type: 'tuple',
-      },
-      { internalType: 'bytes', name: '_mintDetails', type: 'bytes' },
-    ],
-    name: 'setMintDetails',
-    outputs: [],
-    stateMutability: 'nonpayable',
-    type: 'function',
-  },
-  {
-    inputs: [
-      { internalType: 'address', name: '', type: 'address' },
-      { internalType: 'uint256', name: '', type: 'uint256' },
-    ],
-    name: 'signingAuthorities',
-    outputs: [{ internalType: 'address', name: '', type: 'address' }],
-    stateMutability: 'view',
-    type: 'function',
-  },
-  {
-    inputs: [{ internalType: 'address', name: 'newOwner', type: 'address' }],
-    name: 'transferOwnership',
-    outputs: [],
-    stateMutability: 'payable',
-    type: 'function',
-  },
-  { inputs: [], name: 'unpause', outputs: [], stateMutability: 'nonpayable', type: 'function' },
-  {
-    inputs: [{ internalType: 'address', name: '_token', type: 'address' }],
-    name: 'withdraw',
-    outputs: [],
-    stateMutability: 'nonpayable',
-    type: 'function',
-  },
+  }
 ];
